% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{center_curve}
\alias{center_curve}
\title{Centers curves for plotting}
\usage{
center_curve(data_curve)
}
\arguments{
\item{data_curve}{curve data}
}
\value{
a \code{data.frame} with evaluations of the curve
centered at the origin
}
\description{
Centers curves for plotting
}
