% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optimal_t_discrete_closed.R
\name{find_optimal_t_discrete_closed}
\alias{find_optimal_t_discrete_closed}
\title{Finds optimal alignment for discrete closed curves}
\usage{
find_optimal_t_discrete_closed(r, p, s, q, initial_t, eps = 10^-3)
}
\arguments{
\item{r}{time points for p, first is last - 1}

\item{p}{square root velocity vectors, one less than time points in r}

\item{s}{time points for q, first is last - 1}

\item{q}{square root velocity vectors, one less than time points in s}

\item{initial_t}{starting value for the optimization algorithm}

\item{eps}{convergence tolerance}
}
\value{
optimal time points for q, first is last -1
}
\description{
Finds optimal aligned time points for srv curve q to  srv curve p using
coordinate wise optimization.
}
