% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aligned_curves.R
\name{plot.aligned_curves}
\alias{plot.aligned_curves}
\title{Plot method for aligned curves}
\usage{
\method{plot}{aligned_curves}(x, points_col = rainbow, ...)
}
\arguments{
\item{x}{object of class \code{aligned_curves}, usually a result of a call to \code{\link{align_curves}}}

\item{points_col}{which color palette is used for points on the curves,
default is rainbow, see \code{\link{rainbow}} for further options.}

\item{...}{further plotting parameters.}
}
\value{
No value
}
\description{
Plots objects of class \code{aligned_curves}.
Points of same color correspond after the second curve is optimally aligned to the first curve.
}
\seealso{
For examples see documentation of \code{\link{align_curves}}.
}
