% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elastic_mean.R
\name{plot.elastic_mean}
\alias{plot.elastic_mean}
\title{Plot method for planar elastic mean curves}
\usage{
\method{plot}{elastic_mean}(x, asp = 1, col = "red", ...)
}
\arguments{
\item{x}{object of class \code{elastic_mean},
usually a result of a call to \code{\link{compute_elastic_mean}}}

\item{asp}{numeric, giving the aspect ratio of the two coordinates,
see \code{\link{plot.window}} for details.}

\item{col}{color of the mean curve.}

\item{...}{further plotting parameters.}
}
\value{
No value
}
\description{
Plots objects of class \code{elastic_mean}.
}
\seealso{
For examples see documentation of \code{\link{compute_elastic_mean}}.
}
