% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elastic_reg_model.R
\name{plot.elastic_reg_model}
\alias{plot.elastic_reg_model}
\title{Plot method for planar elastic regression models}
\usage{
\method{plot}{elastic_reg_model}(x, asp = 1, col = "red", ...)
}
\arguments{
\item{x}{object of class \code{elastic_reg_model},
usually a result of a call to \code{\link{fit_elastic_regression}}}

\item{asp}{numeric, giving the aspect ratio of the two coordinates,
see \code{\link{plot.window}} for details.}

\item{col}{color of the predicted curves.}

\item{...}{further plotting parameters.}
}
\value{
No value
}
\description{
Plots objects of class \code{elastic_reg_model}.
}
\seealso{
For examples see documentation of \code{\link{fit_elastic_regression}}.
}
