% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_elastic_reg_model.R
\name{predict.elastic_reg_model}
\alias{predict.elastic_reg_model}
\title{Predict method for elastic regression models}
\usage{
\method{predict}{elastic_reg_model}(object, newdata = NULL, t_grid = seq(0, 1, 0.01), ...)
}
\arguments{
\item{object}{object of class \code{elastic_reg_model},
usually a result of a call to \code{\link{fit_elastic_regression}}}

\item{newdata}{an optional \code{data.frame} in which to look for variables with which to predict.
If not given, the fitted values are used.}

\item{t_grid}{grid on which the predicted curves are evaluated.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a \code{list} of \code{data.frame}s with predicted curves
}
\description{
predicted curves for elastic regression model objects.
}
\seealso{
For examples see documentation of \code{\link{fit_elastic_regression}}.
}
