% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun_closed_reg.R
\name{project_curve_on_closed}
\alias{project_curve_on_closed}
\title{Close open curve via projection on derivative level.}
\usage{
project_curve_on_closed(data_curve)
}
\arguments{
\item{data_curve}{\code{data.frame} with values of the curve.}
}
\value{
a \code{data.frame} with closed curve.
}
\description{
Close open curve via projection on derivative level.
}
