% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{srvf_to_curve}
\alias{srvf_to_curve}
\title{Retransform srv curve back to curve}
\usage{
srvf_to_curve(t, srv_curve)
}
\arguments{
\item{t}{time points at which the resulting curve shall be evaluated.}

\item{srv_curve}{srv curve as a function of one parameter,
needs to be vectorised.}
}
\value{
a \code{matrix} with curve evaluations at time points t in its columns,
rows correspond to coordinate directions
}
\description{
Retransform srv curve back to curve
}
